streamlit_app/requirements.txt
streamlit==1.33.0
pillow
requests   # call the FastAPI endpoint
matplotlib
